Initial commit for refactor. The purpose of this refactor is two fold:
1. Move DeGiro into a subdirectory + preserve LICENCE information
2. Convert this repo to be buildable by bazel

The following actions were taken to achieve those goals:
- Create deps/ directory for code dependencies.
- Move all DeGiro code into deps/
- Move LICENSE file in deps/degiro/
- Modify deps/degiro/__init__.py to import libraries from current directory.
- Add bazel BUILD file for DeGiro.
- Add bazel WORKSPACE file.
- Add src/ directory for future commits.
- Add requirements.txt file in src/.
