+++ deps/degiroapi/BUILD
@@ -1,0 +1,7 @@
+load("@rules_python//python:defs.bzl", "py_library")
+
+py_library(
+    name = "degiroapi",
+    srcs = glob(["*.py"]),
+    visibility = ["//visibility:public"],
+)
