+++ deps/degiroapi/datatypes.py
