+++ deps/degiroapi/intervaltypes.py
