+++ deps/degiroapi/product.py
